app/requirements.txt
flake8==3.7.9
autopep8==1.5.4
pylint
python-dotenv==0.13.0
passlib[bcrypt] 
pyjwt==1.7.1
pytest
pytest-cov
pytest-asyncio
pytest-trio
pytest-tornasync
peewee_async==0.6.3a0
psycopg2==2.8.5
aiopg==1.0.0
uvicorn==0.10.8
fastapi==0.45.0
gunicorn
python-multipart
pydantic==1.5.1
email-validator==1.1.1
uvloop==0.14.0